--- .env-cmdrc
@@ -1,17 +1,0 @@
-{
-  "blocks": {
-    "GOBSSTONES_STARTUP_MODE": "blocks",
-    "GOBSSTONES_STARTUP_NAME": "Gobstones Jr",
-    "GOBSSTONES_STARTUP_FILENAME": "gobstones-jr"
-  },
-  "code": {
-    "GOBSSTONES_STARTUP_MODE": "code",
-    "GOBSSTONES_STARTUP_NAME": "Gobstones Sr",
-    "GOBSSTONES_STARTUP_FILENAME": "gobstones-sr"
-  },
-  "teacher": {
-    "GOBSSTONES_STARTUP_MODE": "teacher",
-    "GOBSSTONES_STARTUP_NAME": "Gobstones Teacher",
-    "GOBSSTONES_STARTUP_FILENAME": "gobstones-teacher"
-  }
-}
