--- .idea/jsLibraryMappings.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="JavaScriptLibraryMappings">
-    <file url="file://$PROJECT_DIR$" libraries="{gobstones-web node_modules}" />
-    <file url="PROJECT" libraries="{index.umd}" />
-  </component>
-</project>
