+++ app/config.json
@@ -1,1 +1,1 @@
-{"code-panel-percentage":0.6,"selected-view-mode":0,"language":"es"}
+{"code-panel-percentage":null,"selected-view-mode":0,"language":"es","selected-guide":0}
